better-auth-authentication/backend/pyproject.toml
[project]
name = "tasknest-backend"
version = "0.1.0"
description = "TaskNest Phase 2 - Backend API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlmodel>=0.0.14",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.0",
    "requests>=2.31.0",
    "pytest>=9.0.2",
    "httpx>=0.28.1",
    "aiosqlite>=0.22.1",
    "pyjwt[crypto]>=2.11.0",
    "cryptography>=46.0.4",
    "openai>=2.21.0",
    "grpcio>=1.78.0",
    "grpcio-status>=1.71.2",
    "google-genai>=1.63.0",
    "python-dateutil>=2.8.2",
    "mcp>=1.0.0",
    "apscheduler>=3.11.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "black>=23.12.0",
    "ruff>=0.1.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.hatch.build.targets.wheel]
packages = ["src/app"]
